requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
convertdate==2.4.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
PyMeeus==0.5.12
python-dotenv==1.0.1
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
urllib3==2.2.3
